+++ src/components/Match/MatchHeader/MatchHeader.css
@@ -131,3 +131,83 @@
     }
   }
 }
+
+.overviewHead {
+  & hr {
+    border: 0;
+    height: 1px;
+    background-color: var(--primaryTextColor);
+    background: linear-gradient(to right, var(--primaryTextColor) , rgba(0, 0, 0, 0)); 
+    opacity: .1;
+    margin: 20px 0 25px;
+  }
+  & .matchNumbers {
+    text-align: right;
+    padding-top: 5px;
+    
+    & span {
+      font-size: var(--fontSizeSmall);
+      color: var(--colorMutedLight);
+      margin-left: 20px;
+      white-space: pre-line;
+    }
+    & > div {
+      display: inline;
+      vertical-align: middle;
+
+      &:first-of-type {
+        & span {
+          margin-left: 0;
+        }
+      }
+      & > div {
+        display: inline;
+
+        @media only screen and (max-width: 560px) {
+          display: block;
+          line-height: 2;
+
+          & span {
+            margin-left: 0;
+          }
+        }
+      }
+    }
+  }
+  & .matchButtons {
+    /* Material-ui buttons */
+    & a {
+      float: left;
+      margin-right: 10px !important;
+      line-height: 34px !important;
+    }
+    & a:last-of-type {
+      margin-right: 0 !important;
+    }
+
+    @media only screen and (max-width: 600px) {
+      & a {
+        min-width: 24px !important;
+
+        & span {
+          font-size: 0px !important;
+          padding-left: 0 !important;
+          padding-right: 12px !important;
+        }
+      }
+    }
+  }
+
+  @media only screen and (max-width: 1199px) {
+    & .matchNumbers {
+      text-align: center;
+      margin: 10px 0 5px;
+    }
+    & .matchButtons {
+      & div {
+        display: table;
+        margin: 0 auto;
+      }
+    }
+  }
+}
