+++ src/lang/en.json
@@ -278,7 +278,7 @@
   "th_abilities": "Abilities",
   "th_items": "Items",
   "th_mmr": "MMR",
-  "th_level": "Lvl",
+  "th_level": "LVL",
   "th_kills": "K",
   "th_kills_per_min": "KPM",
   "th_deaths": "D",
